==> workflows/04_calculate_ensemble_metrics.Rmd <==
---
title: "Computing End-of-Century Temperature Metrics"
author: "Joe Brown"
date: "2024-09-18"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Goal

The goal of this markdown file is to continue the analysis by computing end-of-century GSAT anomaly metrics from model results. 

The model results need to be normalized so that end-of-century warming metrics are relative to the 1995-2015 reference period. To do this we will take advantage of a sourced function that transforms data to be relative to a reference year range. 

# Load libaries, helper functions, and data

## Libraries

The libraries that we need for this script are the same as previous steps in the workflow. Will skip over loading these.

## Loading data

For this analysis we will need the initial model results. We will load other data as need.

```{r Load data for scipt}
# model_result <- readRDS("data/model_result.RDS")

```

# Normalize model results to 1995-2014 reference period

```{r calculate global tas relative to 1995-2014}
rel_model_results <- lapply(model_result, function(df) {
  
  # subset to only include global_tas
  data_subset <- subset(df, 
                        variable == "global_tas")
  
  # compute normalize data
  normalized_data <- normalize_to_reference(data_subset,
                                            var = "global_tas",
                                            ref_start = 1995,
                                            ref_end = 2014)
  
  # replace values in results with normalized data
  data_subset$value <- normalized_data
  
  return(data_subset)
})

```

We now have the `rel_model_results`, the `value` column in each data frame has been replaced with `global_tas` values relative to the specified reference period.

# Initialize Metric 

We are initializing a metric to compute the end-of-century (`2081:2100`) median warming (`global_tas`).

```{r, initializing warming metric}
long_warming_metric <- new_metric(var = 'global_tas', years = 2081:2100, op = median)
mid_warming_metric <- new_metric(var = 'global_tas', years = 2041:2060, op = median)
short_warming_metric <- new_metric(var = 'global_tas', years = 2021:2040, op = median)

metric_list <- list('long' = long_warming_metric, 
                    'mid' = mid_warming_metric, 
                    'short' = short_warming_metric)
```

# Compute metrics 

Using the `warming_metric` and `rel_model_results` we compute the end-of-century warming metrics for each ensemble member.

```{r compute wamring metrics for each ensemble memeber}
warming_results <- lapply(rel_model_results, function(df){
  
  # setNames of the metric elements to match the names of the metrics used for calculation
  # loop across each element of the rel_model_results and use each metric in metric list to compute metric results 
  metric_result = setNames(lapply(names(metric_list), function(metric_name) {
    
    # store each metric object for current iteration
    metric = metric_list[[metric_name]]
    
    # use metric and model data of current iteration to compute metrics
    result = metric_calc(df, metric)
    
    # Add a column indicating the name of the metric (for binding later)
    result$term = metric_name
    
    return(result)
    
  }), names(metric_list)) # specify what names should be used to name the elements of the result
    
  return(metric_result)
  
})

```

Save the warming results:

```{r save warming metric results}
saveRDS(warming_results, "data/warming_results.RDS")

```
