==> workflows/05_merge_weights_and_metrics.Rmd <==
---
title: "Merging data for weighted metrics"
author: "Joe Brown"
date: "2024-09-19"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Goal 

Here we want to merge data to get a list of data frames with weighted metrics.

We want the resulting object to be a list of lists (1-219) where each element is a list of SSP scenario data frames containing metric values and weights associated with each. The weights were those found using different combinations of criteria weights. Therefore, the result should be a list (of 219) where each element is a list of SSP scenarios with the following columns: `run_number`, `metric_result`, and `mc_weight`. For each list (1-219) the weights were calculated using a different combination of criteria weights; which can be found in the `criteria_wts` data file.

## Load data

Fix this, don't need doubling data and won't have to if naming is consistent

```{r load data}
# # weight_list = 
# # weight_list <- readRDS("data/model_wts.RDS")
# metric_list <- readRDS("data/warming_results.RDS")
# criteria_weights <- readRDS("data/criteria_wts.RDS")

```

# Merge metrics and weights

For this step we need to merge `weight_list` with `metric_list` by the `run_number` column in the data frames. 

To do this we use the `merge_dfs()` function that we wrote in `helper_functions.R`.

```{r merge weights and metric values}
# Merge constrained weights with warming results
weighted_metric_list <- lapply(constrained_weights, function(criterion_level) {
  
  # Iterate over scenarios in both warming_results and the current criterion level
  merged_scenarios <- Map(function(scenario_weights, scenario_results) {
    
    # Iterate over term lengths (data frames within each scenario)
    lapply(scenario_results, function(term_length_df) {
      
      # Merge term length results with the corresponding weights
      merged <- merge(term_length_df, scenario_weights, by = "run_number")
      return(merged)
    })
    
  }, criterion_level, warming_results) # Align the criterion level with warming_results by scenario
  
  return(merged_scenarios)
})


```

```{r, merge criteria combinations and add scenario column}
# # Use Map to merge the criteria_weights with weighted_metric_list
# weighted_scenario_metrics <- Map(function(element, criteria) {
#   
#   data <- lapply(names(element), function(name) {
#     
#     # copy data
#     df <- element[[name]]
#     
#     # Add the criteria weights as new columns
#     df$temp_wt <- criteria$temp_wt
#     df$co2_wt <- criteria$co2_wt
#     df$ocean_uptake_wt <- criteria$ocean_uptake_wt
#     
#     # add scenario names column
#     df$scenario <- name
#     
#     return(df)
#   
#     })
#   
#   # combine data 
#   return_data <- do.call(rbind, data)
#   
#   # Return the combined data frame
#   return(return_data)
#   
# }, weighted_metric_list, criteria_weights_list)
# 
# # combine all results
# combined_result <- do.call(rbind, weighted_scenario_metrics)
# 
# # split by scenario
# scenario_metrics <- split(combined_result, combined_result$scenario)

```

```{r save the new combined data}
#saveRDS(weighted_scenario_metrics, "data/scenario_metrics.RDS")
```
