==> workflows/06_computing_summary_statistics.Rmd <==
---
title: "Computing Summary statistic and error terms"
author: "Joe Brown"
date: "2024-09-19"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Goal

First computing summary statistics for the temperature metrics and then using IPCC values to compute error terms for each criteria weight combination.

*Note: IPCC values referenced here are stored in the setup.R file* 

# Compute summary statistics

Use `split_data_frame` to compute summary statistics (weighted median and percentiles) for temperature metrics from each SSP scenario. 

```{r}
stat_summary <- lapply(weighted_metric_list, function(criterion_level) {
  
  lapply(criterion_level, function(scenario_level) {
    
    lapply(scenario_level, function(term_df){
      
    lower = quantile(term_df$metric_result, probs = 0.05)
    median = quantile(term_df$metric_result, probs = 0.5)
    upper = quantile(term_df$metric_result, probs = 0.95)
    
    stats = data.frame(
      lower = lower,
      median = median, 
      upper = upper)
    
    return(stats)
    })
  })
})
```

## Compute error terms

Needs notes and documentation.

```{r}
error_results <- lapply(names(stat_summary), function(criterion_name) {
  
  lapply(names(stat_summary[[criterion_name]]), function(scenario_name) {
    
    lapply(names(stat_summary[[criterion_name]][[scenario_name]]), function(term_name) {
      
      # Retrieve the statistics for this term length
      stat_data <- stat_summary[[criterion_name]][[scenario_name]][[term_name]]
      
      # Retrieve the corresponding IPCC values for this scenario and term length
      ipcc_values_for_scenario <- ipcc_term_values[[scenario_name]][[term_name]]
      
      # Calculate the errors for each statistic (median, lower, upper)
      median_error <- compute_error(stat_data$median, ipcc_values_for_scenario[1])
      lower_error  <- compute_error(stat_data$lower, ipcc_values_for_scenario[2])
      upper_error  <- compute_error(stat_data$upper, ipcc_values_for_scenario[3])
      
      # Return error results as a named list or data frame
      error_data <- data.frame(
        criterion = criterion_name,
        scenario = scenario_name,
        term_length = term_name,
        median_error = median_error$percentage_error,
        lower_error = lower_error$percentage_error,
        upper_error = upper_error$percentage_error,
        cumulative_error = median_error$percentage_error + 
                           lower_error$percentage_error + 
                           upper_error$percentage_error
      )
      
      return(error_data)
    })
  })
})

# Flatten the nested list structure into a single data frame
flat_error_results <- do.call(rbind, lapply(error_results, function(criterion_level) {
  do.call(rbind, lapply(criterion_level, function(scenario_level) {
    do.call(rbind, scenario_level)
  }))
}))

# Optionally, reset row names if desired
rownames(flat_error_results) <- NULL

# # Compute error results using lapply
# error_results <- lapply(names(stat_summary), function(scenario_name) {
#   
#   scenario <- stat_summary[[scenario_name]]  # Get metrics for the scenario
#   
#   # Use lapply for criteria weight combinations
#   error_by_criteria_id <- lapply(names(scenario), function(criteria_id) {
#     criteria_data <- scenario[[criteria_id]]  # Get the data frame for the criteria
# 
#     # Use lapply for term lengths and calculate errors
#     error_by_term <- lapply(names(criteria_data), function(term_length) {
#       metric_stats <- criteria_data[[term_length]]  # Get the metric stats for the term length
#       
#       # Retrieve corresponding IPCC values for this scenario and term length
#       ipcc_values_for_scenario <- ipcc_term_values[[scenario_name]][[term_length]]
#       
#       # Calculate the errors for each statistic (median, lower, upper)
#       median_error <- compute_error(metric_stats$median, ipcc_values_for_scenario[1])
#       lower_error  <- compute_error(metric_stats$lower, ipcc_values_for_scenario[2])
#       upper_error  <- compute_error(metric_stats$upper, ipcc_values_for_scenario[3])
#       
#       # Return error results as a data frame
#       data.frame(
#         criteria_id = criteria_id,
#         term_length = term_length,
#         median_error = median_error$percentage_error,  # Adjust if you want to include all error types
#         lower_error = lower_error$percentage_error,
#         upper_error = upper_error$percentage_error, 
#         cumulative_error = median_error$percentage_error + lower_error$percentage_error + upper_error$percentage_error
#       )
#     })
#     
#     # Combine results for each term length into one data frame
#     bind_error <- do.call(rbind, error_by_term)
#     
#     return(bind_error)  # Return the combined data frame for the criteria
#   })
#   
#   # Combine all criteria results for the scenario into one data frame
#   bind_criteria_error <- do.call(rbind, error_by_criteria_id)
# 
#   return(bind_criteria_error)  # Return the combined data frame for the scenario
#   
# })
# 
# 
# # Set top level names of list to match IPCC naming
# names(error_results) <- ssp_names

```

```{r}
# Step 1: Filter the data based on your error criteria
filtered_errors <- flat_error_results %>%
  filter(median_error < 5, lower_error < 10, upper_error < 10)

# Step 2: For each SSP-term_length combination, select the row with the lowest cumulative error
# If there are no records after filtering, we will use the row with the lowest cumulative_error
best_errors <- flat_error_results %>%
  group_by(scenario, term_length) %>%
  # First try to pick the best records based on filtering criteria
  mutate(filtered = scenario %in% filtered_errors$scenario & term_length %in% filtered_errors$term_length) %>%
  filter(filtered | (is.na(filtered) & !duplicated(scenario) & !duplicated(term_length))) %>%
  slice_min(order_by = cumulative_error, n = 1, with_ties = FALSE) %>%
  ungroup()

# Check if all combinations are covered
missing_combinations <- setdiff(paste(flat_error_results$scenario, flat_error_results$term_length), 
                                paste(best_errors$scenario, best_errors$term_length))

if(length(missing_combinations) > 0) {
  # if any combinations are missing, take the row with the lowest cumulative error for those combinations
  print("Some combinations were not fully filtered, here are the combinations missing")
  print(missing_combinations)
} else {
  print("All SSP-term combinations are present in the filtered data")
}

# Display the best errors
print(best_errors)



```


## Putting things together

Where I am at now is needing to put thins together. Ideally I think the best way is a data frame with the following column:

1. `scenario` - name of the SSP scenario.
2. `temp_wt` - the weight values given to observed temperature criterion. 
3. `co2_wt` - the weight values given to observed co2 criterion.
4. `ocean_uptake_wt` - the weight values given to observed ocean C uptake criterion.
5. `metric_stat_values` - the median, lower, and upper bounds from the Hector simulations. Nested in stat_summary (not sure how immediately critical these values are - don't think they are needed for ternary plot).
6. `absolute_error` - the absolute error between the metric estimates from Hector/Matilda and IPCC values. Should be one for each criterion weight combination.
7. `percentage_error` - represents the percent error rather than absolute error.

Where are these pieces of data currently?:

1. `scenario` - in ssp_names
2. `temp_wt` - in `criteria_weights` 
3. `co2_wt` - in `criteria_weights` 
4. `ocean_uptake_wt` - in `criteria_weights` 
5. `metric_stat_values` - skipping this for now -- will need it for another figure later.
6. `absolute_error` - in `error_results`
7. `percentage_error` - in `error_results`


```{r}
# building data frame for each SSP to plot

plotting_list <- lapply(names(error_results), function(df_name) {
  
  data = error_results[[df_name]]
  
  data$temp_wt = rep(criteria_weights$temp_wt, each = 3)
  data$CO2_wt = rep(criteria_weights$co2_wt, each = 3)
  data$ocean_uptake_wt = rep(criteria_weights$ocean_uptake_wt, each = 3)

  return(data)  
})

# Set top level names of list to match IPCC naming
names(plotting_list) <- ssp_names
```

Save plotting data:
```{r}
# save plot data 
saveRDS(plotting_list, "data/plotting_list.rds")
```
